requirements.txt
pydantic==2.5.1
fastapi==0.104.1
SQLAlchemy==2.0.23
python-dotenv==1.0.0
asyncpg==0.29.0
